+++ TopNews/TopNewsiOS/NewsUI/Composers/NewsUIComposer.swift
@@ -32,26 +32,6 @@ private extension NewsViewController {
     }
 }
 
-private final class WeakRefVirtualProxy<T: AnyObject> {
-    private weak var object: T?
-    
-    init(_ object: T) {
-        self.object = object
-    }
-}
-
-extension WeakRefVirtualProxy: NewsLoadingView where T: NewsLoadingView {
-    func display(_ viewModel: NewsLoadingViewModel) {
-        object?.display(viewModel)
-    }
-}
-
-extension WeakRefVirtualProxy: NewsImageView where T: NewsImageView, T.Image == UIImage {
-    func display(_ model: NewsImageViewModel<UIImage>) {
-        object?.display(model)
-    }
-}
-
 private final class NewsViewAdapter: NewsView {
     private weak var controller: NewsViewController?
     private let imageLoader: NewsImageDataLoader
